pyproject.toml
[project]
name = "code-interpreter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.9.0",
    "isort>=7.0.0",
    "langchain>=1.0.1",
    "langchain-experimental>=0.0.42",
    "langchain-ollama>=1.0.0",
    "langchainhub>=0.1.21",
    "python-dotenv>=1.1.1",
    "qrcode>=8.2",
]
